Move side navigation components into their own directory
The property directory is for property editing components... whereas these components are for navigation
