+++ src/components/editor/leftNav/PanelResourceNav.jsx
